# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GYM API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": ""
        }
    ],
...
    "paths": {
...
        "/api/users/{id}":{
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Eliminar un usuario",
                "description": "Eliminar un usuario",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    }
}